+++ R/gene_locus_list.R
@@ -7,6 +7,8 @@
 #' (e.g. GWAS, eQTL, tQTL).
 #' @param topSNPs Output of the function \link[echodata]{import_topSNPs}, 
 #' containing the top (index) SNPs per locus.
+#' @param drop_missing_loci Drop \code{loci} that are missing from 
+#' \code{topSNPs$Locus}.
 #' @param verbose Print messages.
 #' 
 #' @returns Named list
@@ -17,9 +19,13 @@
 #'                         dataset_type = "QTL")
 gene_locus_list <- function(loci = NULL,
                             topSNPs = NULL,
+                            drop_missing_loci = TRUE,
                             dataset_type = NULL,
                             verbose = TRUE){
-    is_qtl <- grepl("qtl$",dataset_type,ignore.case = TRUE)
+    
+    is_qtl <- grepl(pattern = "qtl$",
+                    x = dataset_type,
+                    ignore.case = TRUE)
     if(is.null(loci)) {
         if(any(is.null(c(topSNPs,dataset_type)))){
             stp <- paste(
@@ -35,11 +41,22 @@ gene_locus_list <- function(loci = NULL,
             loci <- unique(topSNPs$Locus)
         } 
     }  
+    #### drop_missing_loci ####
+    if(isTRUE(drop_missing_loci)){
+        missing_loci <- loci[!loci %in% topSNPs$Locus]
+        if(length(missing_loci)>0){
+            messager("Dropping",formatC(length(missing_loci),big.mark = ","),
+                     "loci not present in topSNPs:",
+                     paste("\n -",missing_loci,collapse = ""),
+                     v=verbose)
+            loci <- loci[loci %in% topSNPs$Locus]
+        } 
+    }
     #### Ensure all locus names are unique ####
     if(detect_genes(loci = loci)){
         if((length(unique(loci))!=length(loci)) | 
            all(names(loci)==unname(loci)) ){
-            messager("Reassigning gene-specific locus names",v=verbose)
+            messager("Reassigning gene-specific locus names.",v=verbose)
             loci <- stats::setNames(paste(unname(loci),names(loci),sep="_"),
                                     names(loci))
         }